requirements/base.txt
django==2.1.3
djangorestframework==3.9.0
django-rest-auth[with_social]==0.9.3
djangorestframework-jwt==1.11.0
django-extended-choices==1.3
